fs: move ml.Config to fs package
